Run boot twice to cache
Not sure if/when this changed, but running `boot` the first time on a new system grabs the necessary `boot` binaries, but then ignores subsequent commands. So in order to actually cache REPL and web deps as the comment here indicates, `boot` should be run twice: once to download `boot` proper, and once to actually run the tasks and cache deps.